final refactor of the code
